# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Linka App Api Documentation",
        "description": "This Linka app will provide access to the database and coonsume data to your application",
        "contact": {
            "email": "ganzatambaheritier@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "17e6c4938b676a872ec4c274fc7b165c",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "419": {
                        "description": "Page expired"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "cf6cdd4140fe406a11dc479250e3b8e7",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "419": {
                        "description": "Page expired"
                    }
                }
            }
        },
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get all  User Profiles",
                "operationId": "43d12304c3e621ecce8cc9ccd4b3738e",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/profile/create": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Create new Profile",
                "operationId": "2f8af03be69fcf3005a7868e93c00f9d",
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "firstname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "lastname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nickName",
                        "in": "query",
                        "description": "nickname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "age",
                        "in": "query",
                        "description": "age",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "gender",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "country",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "weight",
                        "in": "query",
                        "description": "weight",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "height",
                        "in": "query",
                        "description": "height",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "personalInfo",
                        "in": "query",
                        "description": "personalInfo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sexualOrientation",
                        "in": "query",
                        "description": "sexualOrientation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lookingFor",
                        "in": "query",
                        "description": "lookingFor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lookingDescription",
                        "in": "query",
                        "description": "lookingDescription",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "profileImagePath",
                        "in": "query",
                        "description": "user Profile Image",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "updated"
                    },
                    "401": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
        "/api/tips": {
            "get": {
                "tags": [
                    "Tips"
                ],
                "summary": "Get All Tips",
                "operationId": "8767c3c692b0a5b735c50771fd0aab99",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/tip/create": {
            "post": {
                "tags": [
                    "Tips"
                ],
                "summary": "Create new tips",
                "operationId": "f5b2ce201b0e94946176f8a3424923c0",
                "parameters": [
                    {
                        "name": "tip_title",
                        "in": "query",
                        "description": "Tip title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description about tips",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Image",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "updated"
                    },
                    "401": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
        "/api/tip/{tip}/update": {
            "put": {
                "tags": [
                    "Tips"
                ],
                "summary": "Update tips new tips",
                "operationId": "90381be9641412dbd3325737cf84a690",
                "parameters": [
                    {
                        "name": "tip_title",
                        "in": "query",
                        "description": "Tip title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "update tips tips",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "description about tips",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated"
                    },
                    "402": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
        "/api/tip/{tip}/get": {
            "get": {
                "tags": [
                    "Tips"
                ],
                "summary": "Get Tip details",
                "operationId": "3f872668c421d7db1be321cb9035d0dd",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/tip/{tip}/delete": {
            "delete": {
                "tags": [
                    "Tips"
                ],
                "summary": "Delete",
                "operationId": "0dd6362290b5f23b0a6fcd70bbd2fde0",
                "responses": {
                    "200": {
                        "description": "deleted"
                    },
                    "401": {
                        "description": "UnAuthorized"
                    }
                }
            }
        }
    }
}